requirements.txt
# Core FastAPI and Web Server
fastapi==0.115.12
uvicorn==0.34.3
pydantic==2.11.7
pydantic-settings==2.9.1

# LangGraph and LangChain
langgraph==0.2.45
langchain==0.3.25
langchain-community==0.3.25
langchain-core==0.3.65
langchain-openai==0.3.23

# Vector Store (FAISS)
faiss-cpu==1.8.0.post1

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.41

# OpenAI
openai==1.86.0

# Environment and Utilities
python-dotenv==1.1.0

# HTTP Client (for health checks)
httpx==0.28.1

